+++ Example-OSX/Common/Components/LayoutExampleCollectionViewItem.xib
@@ -13,9 +13,50 @@
         <customObject id="-1" userLabel="First Responder" customClass="FirstResponder"/>
         <customObject id="-3" userLabel="Application" customClass="NSObject"/>
         <customView identifier="LayoutExampleCollectionViewItem" id="Hz6-mo-xeY">
-            <rect key="frame" x="0.0" y="0.0" width="387" height="204"/>
+            <rect key="frame" x="0.0" y="0.0" width="387" height="95"/>
             <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMinY="YES"/>
-            <point key="canvasLocation" x="93.5" y="237"/>
+            <subviews>
+                <imageView horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="WM9-vm-CJn">
+                    <rect key="frame" x="20" y="20" width="55" height="55"/>
+                    <constraints>
+                        <constraint firstAttribute="width" constant="55" id="Tue-kv-jDB"/>
+                        <constraint firstAttribute="height" constant="55" id="l9z-9Z-0fE"/>
+                    </constraints>
+                    <imageCell key="cell" refusesFirstResponder="YES" alignment="left" imageScaling="proportionallyDown" image="Blueprints-icon-single" id="fA7-B3-mTo"/>
+                </imageView>
+                <textField horizontalHuggingPriority="251" verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="dnH-sE-mZg">
+                    <rect key="frame" x="83" y="58" width="286" height="17"/>
+                    <textFieldCell key="cell" lineBreakMode="clipping" title="Title" id="oSD-Cc-RFR">
+                        <font key="font" metaFont="system"/>
+                        <color key="textColor" name="labelColor" catalog="System" colorSpace="catalog"/>
+                        <color key="backgroundColor" name="textBackgroundColor" catalog="System" colorSpace="catalog"/>
+                    </textFieldCell>
+                </textField>
+                <textField horizontalHuggingPriority="251" verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="fQv-fh-k4V">
+                    <rect key="frame" x="83" y="33" width="286" height="17"/>
+                    <textFieldCell key="cell" lineBreakMode="truncatingTail" title="Label" id="FZS-5b-9Zn">
+                        <font key="font" metaFont="system"/>
+                        <color key="textColor" name="labelColor" catalog="System" colorSpace="catalog"/>
+                        <color key="backgroundColor" name="textBackgroundColor" catalog="System" colorSpace="catalog"/>
+                    </textFieldCell>
+                </textField>
+            </subviews>
+            <constraints>
+                <constraint firstItem="WM9-vm-CJn" firstAttribute="leading" secondItem="Hz6-mo-xeY" secondAttribute="leading" constant="20" id="4fK-Kt-Dd6"/>
+                <constraint firstItem="fQv-fh-k4V" firstAttribute="leading" secondItem="dnH-sE-mZg" secondAttribute="leading" id="Bt8-hq-Hu3"/>
+                <constraint firstAttribute="trailing" secondItem="dnH-sE-mZg" secondAttribute="trailing" constant="20" id="GeB-Un-dUe"/>
+                <constraint firstAttribute="bottom" relation="greaterThanOrEqual" secondItem="WM9-vm-CJn" secondAttribute="bottom" constant="20" id="Wwm-Vg-AKZ"/>
+                <constraint firstItem="dnH-sE-mZg" firstAttribute="top" secondItem="WM9-vm-CJn" secondAttribute="top" id="a2J-Jh-M6y"/>
+                <constraint firstItem="WM9-vm-CJn" firstAttribute="top" secondItem="Hz6-mo-xeY" secondAttribute="top" constant="20" id="aT2-JS-Rsw"/>
+                <constraint firstItem="fQv-fh-k4V" firstAttribute="top" secondItem="dnH-sE-mZg" secondAttribute="bottom" constant="8" id="ccW-Zp-Ygf"/>
+                <constraint firstItem="dnH-sE-mZg" firstAttribute="leading" secondItem="WM9-vm-CJn" secondAttribute="trailing" constant="10" id="kzq-v2-8T3"/>
+                <constraint firstItem="fQv-fh-k4V" firstAttribute="trailing" secondItem="dnH-sE-mZg" secondAttribute="trailing" id="raZ-Be-DLq"/>
+                <constraint firstAttribute="bottom" relation="greaterThanOrEqual" secondItem="fQv-fh-k4V" secondAttribute="bottom" constant="20" id="uZs-eL-KGz"/>
+            </constraints>
+            <point key="canvasLocation" x="93.5" y="199.5"/>
         </customView>
     </objects>
+    <resources>
+        <image name="Blueprints-icon-single" width="385" height="386"/>
+    </resources>
 </document>
