+++ Example-OSX/Resources/Assets.xcassets/Graphics/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  }
+}
